added args, added checks to see if output files exist and skip re-running for those outputs

+++ scripts/build_WGS_species_tree.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import glob
+import warnings
 from tqdm import tqdm
 import subprocess as sp
 from Bio.Nexus import Nexus
@@ -46,6 +47,7 @@ mbscriptname = 'WGS_mrbayes_script.txt'
 logfilename = 'WGS_mrbayes_log.txt'
 
 
+
 def getBase(filename,splitchar='.'):
     base = splitchar.join(os.path.basename(filename).split(splitchar)[0:-1])
     return base
@@ -54,6 +56,8 @@ def multiAlignGenomes(genomde_dir,threads,method='parsnp'):
     #align core genomes of all taxa using parsnp, creates a multi-alingment including all taxas used to build a species tree
     # uses more information thatn 16S genes only and works even when there are not enough annotated 16S genes
     outfile = '{}/WGS.xmfa'.format(base_dir)
+    if os.path.isfile(outfile):
+        return outfile
     if method == 'parsnp':
         # -c uses all files in the target dir
         #-r ! randomly choose a reference fasta
@@ -87,19 +91,22 @@ def trimXMFA(xmfa):
 def convertXMFAToFastas(xmfa):
     #take xmfa and convert each section into a seperate fasta aln file
     #ignore biopython comparison warning
+    #warnings.filterwarnings("ignore", category=BiopythonWarning)
     headers = getHeaders(xmfa)
     trimmed_xmfa = trimXMFA(xmfa)
     records = list(SeqIO.parse(open(trimmed_xmfa),'fasta'))
     n_headers = len(headers)
     for ri in range(0,len(records)//n_headers):
+        base_name = '.'.join(os.path.basename(xmfa).split('.')[0:-1])
+        output_file = '{}/fasta/{}_{}.aln_fna'.format(base_dir,base_name,ri)
+        if os.path.isfile(output_file):
+            next
         aln = records[n_headers*ri:n_headers*(ri+1)]
         seqs = [a.seq.upper() for a in aln]
         if len(set(seqs)) == 1: #if all seqs are the same, no point in including this genome section
             for i in range(0,n_headers):
                 aln[i].id = headers[i].strip('\'\">')
                 aln[i].seq = aln[i].seq.upper()
-            base_name = '.'.join(os.path.basename(xmfa).split('.')[0:-1])
-            output_file = '{}/fasta/{}_{}.aln_fna'.format(base_dir,base_name,ri)
             with open(output_file,'w') as output:
                 SeqIO.write(aln,output,'fasta')
     return None
@@ -116,15 +123,19 @@ def convertFastaToNexus(fasta_file):
 def concatNexusAlignments(processes):
     #take the list of fasta alingments and convert each to a nexus file and concat all the nexus files into 1 alingment
     pool = ThreadPool(processes)
+    already_done = [x.split('.')[0] for x in os.listdir('{}/nexus'.format(base_dir))]
     fastas = ['{}/fasta/{}'.format(base_dir,file) for
-              file in os.listdir('{}/fasta'.format(base_dir))]
+              file in os.listdir('{}/fasta'.format(base_dir))
+              if file.split('.')[0] not in already_done]
     list(tqdm(pool.imap(convertFastaToNexus,fastas),
               total=len(fastas),desc='Fastas to Nexus...'))
+    combined_nexus = '{}/WGS.nex'.format(base_dir)
+    if os.path.isfile(combined_nexus):
+        return combined_nexus
     nexus = ['{}/nexus/{}'.format(base_dir,file) for
               file in os.listdir('{}/nexus'.format(base_dir))]
     nexus = [(filename, Nexus.Nexus(filename)) for filename in nexus]
     combined = Nexus.combine(nexus)
-    combined_nexus = '{}/WGS.nex'.format(base_dir)
     combined.write_nexus_data(filename=open(combined_nexus,'w'))
     return combined_nexus
 
@@ -149,7 +160,6 @@ quit""".format(nexus_alignment)
         os.rename(fp,os.path.join(base_dir,fp))
     return None
 
-
 def main(genus,genome_dir,threads):
     os.system('mkdir -p {}'.format(base_dir))
     os.system('mkdir -p {}/fasta'.format(base_dir))
@@ -157,14 +167,25 @@ def main(genus,genome_dir,threads):
     xmfa = multiAlignGenomes(genome_dir,threads)
     convertXMFAToFastas(xmfa)
     nexus = concatNexusAlignments(threads)
+    if os.path.isfile('{}/WGS_species_tree.con.tree'.format(base_dir)):
+        return None
     buildSpeciesTree(nexus)
     return None
 
 
 if __name__ == '__main__':
-    genus = sys.argv[1]
-    genome_dir = sys.argv[2]
-    threads = int(sys.argv[3])
+    if len(sys.argv) > 1:
+        genus = sys.argv[1]
+    else:
+        genus = os.path.basename(os.getcwd())
+    if len(sys.argv) > 2:
+        genome_dir = sys.argv[2]
+    else:
+        genome_dir = './genome'
+    if len(sys.argv) > 3 :
+        threads = int(sys.argv[3])
+    else:
+        threads = 1
     main(genus,genome_dir,threads)
